+++ packages/fiori-app-sub-generator/test/int/fiori-elements/expected-output/alp_v4/webapp/test/integration/opaTests.qunit.js
@@ -22,7 +22,9 @@ sap.ui.require(
     "sap/ui/thirdparty/qunit-2",
     "sap/ui/qunit/qunit-junit",
     "sap/ui/qunit/qunit-coverage",
-    "testNameSpace/alpv4/test/integration/FirstJourney",
+    "testNameSpace/alpv4/test/integration/SalesOrderItemListJourney.gen",
+    "testNameSpace/alpv4/test/integration/SalesOrderItemObjectPageJourney.gen",
+    "testNameSpace/alpv4/test/integration/MaterialDetailsObjectPageJourney.gen",
 ], function (QUnit) {
     "use strict";
     QUnit.start();